+++ win/CS/HandBrakeWPF/Services/UserSettingService.cs
@@ -338,6 +338,9 @@ namespace HandBrakeWPF.Services
             defaults.Add(UserSettingConstants.SaveLogWithVideo, false);
             defaults.Add(UserSettingConstants.ClearOldLogs, true);
 
+            defaults.Add(UserSettingConstants.ExcludedExtensions, new List<string> { "png", "jpg", "srt", "ass", "ssa", "txt" });
+            defaults.Add(UserSettingConstants.RecursiveFolderScan, false);
+            
             // Preview
             defaults.Add(UserSettingConstants.LastPreviewDuration, 30);
             defaults.Add(UserSettingConstants.UseExternalPlayer, false);
@@ -356,7 +359,6 @@ namespace HandBrakeWPF.Services
             defaults.Add(UserSettingConstants.IsUpdateAvailableBuild, 0);
             defaults.Add(UserSettingConstants.ExtendedQueueDisplay, true);
             defaults.Add(UserSettingConstants.HardwareDetectTimeoutSeconds, 12);
-            defaults.Add(UserSettingConstants.ExcludedExtensions, new List<string> { "png", "jpg", "srt", "ass", "ssa", "txt" });
 
 
             return defaults;