+++ win/CS/HandBrakeWPF/UserSettingConstants.cs
@@ -94,5 +94,6 @@ namespace HandBrakeWPF
         public const string HardwareDetectTimeoutSeconds = "HardwareDetectTimeoutSeconds";
         public const string ShowPresetDesc = "ShowPresetDescription";
         public const string ExcludedExtensions = "ExcludedFileExtensions";
+        public static string RecursiveFolderScan = "RecursiveFolderScan";
     }
 }